feat: convert test to vitest test and run in CI tests
